+++ lib/user.php
@@ -120,11 +120,11 @@ class OC_User {
 	 * setup the configured backends in config.php
 	 */
 	public static function setupBackends() {
-		$backends=OC_Config::getValue('user_backends',array());
+		$backends=OC_Config::getValue('user_backends', array());
 		foreach($backends as $i=>$config) {
 			$class=$config['class'];
 			$arguments=$config['arguments'];
-			if(class_exists($class) and array_search($i,self::$_setupedBackends)===false) {
+			if(class_exists($class) and array_search($i, self::$_setupedBackends)===false) {
 				// make a reflection object
 				$reflectionObj = new ReflectionClass($class);
 
@@ -133,7 +133,7 @@ class OC_User {
 				self::useBackend($backend);
 				$_setupedBackends[]=$i;
 			}else{
-				OC_Log::write('core','User backend '.$class.' not found.',OC_Log::ERROR);
+				OC_Log::write('core', 'User backend '.$class.' not found.', OC_Log::ERROR);
 			}
 		}
 	}
@@ -179,10 +179,10 @@ class OC_User {
 				if(!$backend->implementsActions(OC_USER_BACKEND_CREATE_USER))
 					continue;
 
-				$backend->createUser($uid,$password);
+				$backend->createUser($uid, $password);
 				OC_Hook::emit( "OC_User", "post_createUser", array( "uid" => $uid, "password" => $password ));
 
-				return true;
+				return self::userExists($uid);
 			}
 		}
 		return false;
@@ -204,12 +204,19 @@ class OC_User {
 			foreach(self::$_usedBackends as $backend) {
 				$backend->deleteUser($uid);
 			}
+			if (self::userExists($uid)) {
+				return false;
+			}
 			// We have to delete the user from all groups
 			foreach( OC_Group::getUserGroups( $uid ) as $i ) {
 				OC_Group::removeFromGroup( $uid, $i );
 			}
 			// Delete the user's keys in preferences
 			OC_Preferences::deleteUser($uid);
+
+			// Delete user files in /data/
+			OC_Helper::rmdirr(OC_Config::getValue( "datadirectory", OC::$SERVERROOT."/data" ) . '/'.$uid.'/');
+
 			// Emit and exit
 			OC_Hook::emit( "OC_User", "post_deleteUser", array( "uid" => $uid ));
 			return true;
@@ -325,10 +332,12 @@ class OC_User {
 			foreach(self::$_usedBackends as $backend) {
 				if($backend->implementsActions(OC_USER_BACKEND_SET_PASSWORD)) {
 					if($backend->userExists($uid)) {
-						$success |= $backend->setPassword($uid,$password);
+						$success |= $backend->setPassword($uid, $password);
 					}
 				}
 			}
+			// invalidate all login cookies
+			OC_Preferences::deleteApp($uid, 'login_token');
 			OC_Hook::emit( "OC_User", "post_setPassword", array( "uid" => $uid, "password" => $password ));
 			return $success;
 		}
@@ -363,8 +372,7 @@ class OC_User {
 	 * @param $password The password
 	 * @returns string
 	 *
-	 * Check if the password is correct without logging in the user
-	 * returns the user id or false
+	 * returns the path to the users home directory
 	 */
 	public static function getHome($uid) {
 		foreach(self::$_usedBackends as $backend) {
@@ -472,9 +480,10 @@ class OC_User {
 	 */
 	public static function setMagicInCookie($username, $token) {
 		$secure_cookie = OC_Config::getValue("forcessl", false);
-		setcookie("oc_username", $username, time()+60*60*24*15, '', '', $secure_cookie);
-		setcookie("oc_token", $token, time()+60*60*24*15, '', '', $secure_cookie);
-		setcookie("oc_remember_login", true, time()+60*60*24*15, '', '', $secure_cookie);
+		$expires = time() + OC_Config::getValue('remember_login_cookie_lifetime', 60*60*24*15);
+		setcookie("oc_username", $username, $expires, '', '', $secure_cookie);
+		setcookie("oc_token", $token, $expires, '', '', $secure_cookie, true);
+		setcookie("oc_remember_login", true, $expires, '', '', $secure_cookie);
 	}
 
 	/**
@@ -484,8 +493,8 @@ class OC_User {
 		unset($_COOKIE["oc_username"]);
 		unset($_COOKIE["oc_token"]);
 		unset($_COOKIE["oc_remember_login"]);
-		setcookie("oc_username", NULL, -1);
-		setcookie("oc_token", NULL, -1);
-		setcookie("oc_remember_login", NULL, -1);
+		setcookie("oc_username", null, -1);
+		setcookie("oc_token", null, -1);
+		setcookie("oc_remember_login", null, -1);
 	}
 }
